requirements.txt
langchain
langchain-community
# llamaapi
langchain_experimental
text_generation
sentence-transformers
faiss-gpu
langchain-mistralai
langgraph
bs4